test(character): 收编重复的 _wav_bytes 辅助函数，补强音频上传参数断言
- _wav_bytes 从 test_audio_utils.py/test_files_router.py 两处重复定义
  收编进 tests/conftest.py 共用
- StudioCanvasRouter.test.tsx 的部分失败保存用例补上对
  API.uploadFile("character_audio_ref", ...) 调用参数的断言，之前只
  验证了失败路径，未验证走的是专用音频上传通道
